simplify code

+++ Bitstream.cs
@@ -12,7 +12,7 @@ using System;
 class Bitstream
 {
 	internal short end, ptr; // was uchar in c
-	internal long sr;
+	internal uint sr;
 	internal int file_bytes;
 	internal int error, bc;
 	internal System.IO.BinaryReader file;
